refactor(string-adder): refactor Number
