+++ src/main/kotlin/adder/model/Numbers.kt
@@ -1,4 +1,4 @@
-package adder
+package adder.model
 
 data class Numbers(val nums: List<Number>) {
     constructor(input :Collection<String>): this(input.map(::Number))
