+++ src/test/kotlin/adder/NumberTest.kt
@@ -1,5 +1,6 @@
 package adder
 
+import adder.model.Number
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
